chore: move DummyRecord to secmet helpers for consistency
